+++ src/views/PaymentView.tsx
@@ -1,6 +1,6 @@
 import Banner from "../Components/Banner";
-import {ICity} from "../interfaces/ICity";
-import {CityAPI} from '../CityAPI';
+//import {ICity} from "../interfaces/ICity";
+//import {CityAPI} from '../CityAPI';
 import './PaymentView.css';
 import React, {useState} from 'react';
 import CardDetails from "../Components/CardDetails.tsx";
@@ -8,11 +8,12 @@ import PhoneDetails from "../Components/PhoneDetails.tsx";
 import GiftDetails from "../Components/GiftDetails.tsx";
 import StatusBar from "../Components/StatusBar.tsx";
 import InvoiceDetails from "../Components/InvoiceDetails.tsx";
+import AddressForm from "../Components/AddressForm.tsx";
 
 
 export function PaymentView() {
 
-    const [city, setCity] = useState("")
+    //const [city, setCity] = useState("")
 
     const [paymentOption, setPaymentOption] = useState('credit_card'); // Default payment option
 
@@ -41,16 +42,16 @@ export function PaymentView() {
 
 
 
-    async function ValidateZip(value: string) {
+    /*async function ValidateZip(value: string) {
         const newCity: ICity[] = await CityAPI.getCity()
         const result = newCity.find((a) => a.nr.toString() === value)
 
         if (result) {
             setCity(result.navn)
         }
-    }
+    }*/
 
-    function DeliveryAddress() {
+    /*function DeliveryAddress() {
         return (
             <fieldset>
                 <legend>Delivery address</legend>
@@ -75,7 +76,7 @@ export function PaymentView() {
                 </div>
             </fieldset>
         )
-    }
+    }*/
 
 
     return (
@@ -100,24 +101,7 @@ export function PaymentView() {
                 <fieldset>
                     <legend>Billing address</legend>
                     <div className="form-control-group">
-                        <label htmlFor="country">Country </label>
-                        <select className="form-control" id="country" name="user_country" autoFocus>
-                            <option selected>Denmark</option>
-                        </select>
-                        <label htmlFor="address1">Street address </label>
-                        <input className="form-control" type="text" id="address1" name="user_address1" required
-                               maxLength={65}/>
-                        <input className="form-control" type="text" id="address2" name="user_address2" maxLength={65}/>
-                        <label htmlFor="zip">Postal code </label>
-                        <input className="form-control" type="text" id="zip" name="user_zip" required
-                               onChange={e => ValidateZip(e.target.value)}
-
-                        />
-                        <label htmlFor="city">City </label>
-                        <input className="form-control" type="text" id="city" name="user_city"
-                               onChange={(e) => setCity(e.target.value)}
-                               value={city}/>
-
+                        <AddressForm/>
                         <div>
                             <input
                                 className="form-control"
@@ -131,9 +115,15 @@ export function PaymentView() {
                         </div>
                     </div>
                 </fieldset>
-
-                {deliveryAddress && <DeliveryAddress/>}
-
+                
+                    {deliveryAddress && 
+                        (<fieldset>
+                            <legend>Delivery address</legend>
+                            <AddressForm/>
+                            </fieldset>
+                        )
+                    }
+                
                 <div>
                     <fieldset>
                         <legend>Payment method</legend>
